+++ portlets/opencps-portlet/docroot/WEB-INF/src/org/opencps/dossiermgt/search/DossierPartSearchTerms.java
@@ -40,7 +40,7 @@ public class DossierPartSearchTerms extends DossierPartDisplayTerms{
 	    required = ParamUtil.getBoolean(portletRequest, DOSSIERPART_REQUIRED);
 	    
 	    partType = ParamUtil.getInteger(portletRequest, DOSSIERPART_PARTTYPE);
-	    sibling = ParamUtil.getInteger(portletRequest, DOSSIERPART_SIBLING);
+	    sibling = ParamUtil.getDouble(portletRequest, DOSSIERPART_SIBLING);
     }
 	
     public long getDossierpartId() {
@@ -163,12 +163,12 @@ public class DossierPartSearchTerms extends DossierPartDisplayTerms{
     	this.partType = partType;
     }
 	
-    public int getSibling() {
+    public double getSibling() {
     
     	return sibling;
     }
 	
-    public void setSibling(int sibling) {
+    public void setSibling(double sibling) {
     
     	this.sibling = sibling;
     }
@@ -190,5 +190,5 @@ public class DossierPartSearchTerms extends DossierPartDisplayTerms{
 	protected boolean required;
 	
 	protected int partType;  
-	protected int sibling;
+	protected double sibling;
 }
